+++ src/main/java/ru/alina/repository/NotationRepository.java
@@ -5,12 +5,12 @@ import ru.alina.model.Notation;
 import java.util.List;
 
 public interface NotationRepository {
-    Notation save(Notation notation, int userId);
+    Notation save(Notation notation, Long userId);
 
-    void delete(int id, int userId);
+    void delete(Long id, Long userId);
 
     // null if meal do not belong to userId
-    Notation get(int id, int userId);
+    Notation get(Long id, Long userId);
 
-    List<Notation> getAll(int userId);
+    List<Notation> getAll(Long userId);
 }
